修正

+++ db/schema.rb
@@ -43,8 +43,8 @@ ActiveRecord::Schema[8.0].define(version: 2025_07_15_014631) do
   end
 
   create_table "favorites", force: :cascade do |t|
-    t.bigint "user_id", null: false
-    t.bigint "post_id", null: false
+    t.integer "user_id", null: false
+    t.integer "post_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["post_id"], name: "index_favorites_on_post_id"
@@ -52,8 +52,8 @@ ActiveRecord::Schema[8.0].define(version: 2025_07_15_014631) do
   end
 
   create_table "post_tags", force: :cascade do |t|
-    t.bigint "post_id", null: false
-    t.bigint "tag_id", null: false
+    t.integer "post_id", null: false
+    t.integer "tag_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
     t.index ["post_id"], name: "index_post_tags_on_post_id"
@@ -63,7 +63,7 @@ ActiveRecord::Schema[8.0].define(version: 2025_07_15_014631) do
   create_table "posts", force: :cascade do |t|
     t.string "title"
     t.text "body"
-    t.bigint "user_id", null: false
+    t.integer "user_id", null: false
     t.string "image"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
